#28 [Fix] 북마크 원복 + 수정 추가

+++ src/main/java/com/mpnp/baechelin/bookmark/service/BookmarkService.java
@@ -32,17 +32,20 @@ import java.util.stream.Collectors;
 public class BookmarkService {
 
     private final BookmarkRepository bookmarkRepository;
-    private final FolderRepository   folderRepository;
-    private final StoreRepository    storeRepository;
-    private final UserRepository     userRepository;
+    private final FolderRepository folderRepository;
+    private final StoreRepository storeRepository;
+    private final UserRepository userRepository;
     private final StoreService storeService;
 
     @Transactional
     public void bookmark(BookmarkRequestDto bookmarkRequestDto, String socialId) {
         // 북마크 폴더 생성하는 Flow를 따르므로 bookmarkRequestDto의 folderId는 존재
-        Folder folder = folderRepository.findById(bookmarkRequestDto.getFolderId()).orElseThrow(()-> new CustomException(ErrorCode.NO_FOLDER_FOUND));
-        Store  store  = storeRepository.findById((long) bookmarkRequestDto.getStoreId()).orElseThrow(()-> new CustomException(ErrorCode.NO_BOOKMARK_FOUND));
-        User   user   = userRepository.findBySocialId(socialId); if(user == null) {throw new CustomException(ErrorCode.NO_USER_FOUND); }
+        Folder folder = folderRepository.findById(bookmarkRequestDto.getFolderId()).orElseThrow(() -> new CustomException(ErrorCode.NO_FOLDER_FOUND));
+        Store store = storeRepository.findById((long) bookmarkRequestDto.getStoreId()).orElseThrow(() -> new CustomException(ErrorCode.NO_BOOKMARK_FOUND));
+        User user = userRepository.findBySocialId(socialId);
+        if (user == null) {
+            throw new CustomException(ErrorCode.NO_USER_FOUND);
+        }
 
         Bookmark bookmark = Bookmark
                 .builder()
@@ -53,17 +56,21 @@ public class BookmarkService {
 
         if (!bookmarkRepository.existsByStoreIdAndUserId(store, user)) {
             bookmarkRepository.save(bookmark);
-            storeService.updateBookmarkCnt(store,socialId);
+            storeService.updateBookmarkCnt(store, socialId);
         }
     }
+
     @Transactional
     public void bookmarkDelete(Long storeId, String socialId) {
-        User user = userRepository.findBySocialId(socialId); if(user == null) { throw new CustomException(ErrorCode.NO_USER_FOUND); }
-        Store store = storeRepository.findById(storeId).orElseThrow(()-> new CustomException(ErrorCode.NO_STORE_FOUND));
+        User user = userRepository.findBySocialId(socialId);
+        if (user == null) {
+            throw new CustomException(ErrorCode.NO_USER_FOUND);
+        }
+        Store store = storeRepository.findById(storeId).orElseThrow(() -> new CustomException(ErrorCode.NO_STORE_FOUND));
         Bookmark bookmark = bookmarkRepository.findByStoreIdAndUserId(store, user).orElseThrow(() -> new CustomException(ErrorCode.NO_BOOKMARK_FOUND));
         store.removeBookmark(bookmark);
         bookmarkRepository.delete(bookmark);
-        storeService.updateBookmarkCnt(store,socialId);
+        storeService.updateBookmarkCnt(store, socialId);
     }
 
 
@@ -74,7 +81,7 @@ public class BookmarkService {
         Page<Bookmark> bookmarkPage = bookmarkRepository.findAllByUserId(user, pageable);
 
         List<BookmarkInfoDto> bookmarkList = new ArrayList<>();
-        for(Bookmark bookmark: bookmarkPage){
+        for (Bookmark bookmark : bookmarkPage) {
             BookmarkInfoDto bookmarkInfoDto = new BookmarkInfoDto(bookmark);
             bookmarkList.add(bookmarkInfoDto);
         }
@@ -87,7 +94,7 @@ public class BookmarkService {
         if (user == null) {
             throw new CustomException(ErrorCode.NO_USER_FOUND);
         }
-        Folder folder = folderRepository.findById(folderId).orElseThrow(()-> new CustomException(ErrorCode.NO_FOLDER_FOUND));
+        Folder folder = folderRepository.findById(folderId).orElseThrow(() -> new CustomException(ErrorCode.NO_FOLDER_FOUND));
         Page<Bookmark> pagedBookmark = bookmarkRepository.findAllByFolderId(folder, pageable);
         return new BookmarkPagedResponseDto(pagedBookmark);
     }
